backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
crewai>=0.1.8
langchain-community>=0.0.10
python-dotenv>=1.0.0
yt-dlp>=2023.11.16
opencv-python>=4.8.1.78
openai-whisper>=20231117